nvidia_parser: Fix failure to parse errors with no file index
This for example happens in a `composite.sh` with:

```glsl
aaaaaa
```

Which will produce:

```
(0) : error C0000: syntax error, unexpected $end, expecting "::" at token "<EOF>"
```

Previously, because the <filepath> match group was required for the
match, these lines would just be skipped over. Now this produces a
nice error pointing to the end of the file.